Creating PopUp component for creating new exercise
